revert(test): address coverage runner review follow-ups
