+++ package.json
@@ -37,7 +37,7 @@
     "coverage": "npm run coverage:clear && npm run coverage:unit && npm run coverage:component && npm run coverage:e2e && npm run coverage:cypress:merge && npm run coverage:merge && npm run coverage:check:baseline && npm run coverage:check:migrated && npm run coverage:report && npm stop",
     "coverage:unit": "vitest run --coverage",
     "precoverage:e2e": "npm run build -- --mode test",
-    "coverage:e2e": "node test/unit/run-cypress-coverage.mjs e2e",
+    "coverage:e2e": "vite preview -m test & cypress run && node test/unit/save-cypress-coverage.mjs e2e",
     "coverage:component": "cypress run --component && node test/unit/save-cypress-coverage.mjs component",
     "coverage:check:baseline": "node test/unit/check-coverage-baseline.mjs",
     "coverage:check:migrated": "node test/unit/check-migrated-coverage.mjs",
