+++ src/js/services/dialer.test.js
@@ -156,12 +156,11 @@ describe('Dialer Service', () => {
       first_name: 'Pat',
       last_name: 'Ient',
     });
-    const resetSpy = vi.spyOn(Backbone.Collection.prototype, 'reset');
 
     service._addPatient(patient);
     service.showPatientLinks(null);
 
-    expect(resetSpy).toHaveBeenCalledWith();
+    expect(five9Init).not.toHaveBeenCalled();
   });
 
   it('adds the action patient when one exists', () => {
